requirements.txt
# =============================================================================
# Automend MLOps Monorepo - Consolidated Requirements
# =============================================================================
# Merged from all 6 dataset pipelines for reproducibility
# Use with conda env: mlops_project (already has DVC, Airflow installed)
# =============================================================================

# === Core Data Processing ===
polars>=1.0.0
pyarrow>=15.0.0
numpy>=1.24.0
requests>=2.31.0

# === Distributed Compute ===
ray[default]>=2.10.0

# === Data Versioning ===
dvc>=3.40.0
dvc-s3>=3.0.1

# === Schema & Validation ===
jsonschema>=4.21.0
pydantic>=2.0

# === ML & Bias Detection ===
fairlearn>=0.9.0
scikit-learn>=1.3.0

# === HuggingFace Integration ===
datasets>=2.18.0
huggingface-hub>=0.22.0

# === Testing ===
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-mock>=3.12.0

# === Utilities ===
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.66.0
loguru>=0.7.0
slack-sdk>=3.27.0
html2text>=2020.1.16

# === Google AI (for Dataset 4 synthetic generation) ===
google-genai>=1.0

# === Legacy (kept for Fairlearn bridge and transition) ===
pandas>=2.0.0

# =============================================================================
# Note: Apache Airflow runs via Docker Compose - see docker-compose.yaml
# For local development without Docker:
# apache-airflow>=2.8.0
# =============================================================================
